+++ python/shipDet_conf.py
@@ -240,6 +240,52 @@ def configure_veto(yaml_file, z0):
 
     detectorList.append(Veto)
 
+def configure_strawtubes(yaml_file, ship_geo):
+    with open(yaml_file) as file:
+        config = yaml.safe_load(file)
+
+    ship_geo.strawtubes_geo = AttrDict(config['SST'])
+
+    # Straw tubes in decay vessel if vacuum, otherwise outside in air
+    ship_geo.strawtubes_geo.medium = "vacuums" if ship_geo.DecayVolumeMedium == "vacuums" else "air"
+    Strawtubes = ROOT.strawtubes(ship_geo.strawtubes_geo.medium)
+    Strawtubes.SetzPositions(
+        ship_geo.strawtubes_geo.z1Position,
+        ship_geo.strawtubes_geo.z2Position,
+        ship_geo.strawtubes_geo.z3Position,
+        ship_geo.strawtubes_geo.z4Position,
+    )
+    Strawtubes.SetApertureArea(
+        ship_geo.strawtubes_geo.width,
+        ship_geo.strawtubes_geo.height,
+    )
+    Strawtubes.SetStrawDiameter(
+        ship_geo.strawtubes_geo.outer_straw_diameter,
+        ship_geo.strawtubes_geo.wall_thickness,
+    )
+    Strawtubes.SetStrawPitch(
+        ship_geo.strawtubes_geo.straw_pitch,
+        ship_geo.strawtubes_geo.y_layer_offset,
+    )
+    Strawtubes.SetDeltazLayer(ship_geo.strawtubes_geo.delta_z_layer)
+    Strawtubes.SetStereoAngle(ship_geo.strawtubes_geo.view_angle)
+    Strawtubes.SetWireThickness(ship_geo.strawtubes_geo.wire_thickness)
+    Strawtubes.SetDeltazView(ship_geo.strawtubes_geo.delta_z_view)
+    Strawtubes.SetFrameMaterial(ship_geo.strawtubes_geo.frame_material)
+    Strawtubes.SetStationEnvelope(
+        ship_geo.strawtubes_geo.station_width,
+        ship_geo.strawtubes_geo.station_height,
+        ship_geo.strawtubes_geo.station_length,
+    )
+
+    #For digitization
+    Strawtubes.SetStrawResolution(
+        ship_geo.strawtubes.v_drift,
+        ship_geo.strawtubes.sigma_spatial,
+    )
+
+    detectorList.append(Strawtubes)
+
 
 def configure(run, ship_geo):
     # ---- for backward compatibility ----
@@ -368,53 +414,10 @@ def configure(run, ship_geo):
         ship_geo.decayVolume.z0,
     )
 
-
-    if ship_geo.strawDesign > 1:
-        # for backward compatibility
-        if ship_geo.strawDesign == 10 and not hasattr(
-            ship_geo.strawtubes, "DeltazFrame"
-        ):
-            ship_geo.strawtubes.DeltazFrame = 2.5 * u.cm
-            ship_geo.strawtubes.FrameLateralWidth = 1.2 * u.m
-            ship_geo.strawtubes.FrameMaterial = "steel"
-        elif not hasattr(ship_geo.strawtubes, "DeltazFrame"):
-            ship_geo.strawtubes.DeltazFrame = 10.0 * u.cm
-            ship_geo.strawtubes.FrameLateralWidth = 1.0 * u.cm
-            ship_geo.strawtubes.FrameMaterial = "aluminium"
-        ship_geo.strawtubes.medium = "vacuums" if ship_geo.DecayVolumeMedium == "vacuums" else "air"
-
-        Strawtubes = ROOT.strawtubes(ship_geo.strawtubes.medium)
-        Strawtubes.SetZpositions(
-            ship_geo.TrackStation1.z,
-            ship_geo.TrackStation2.z,
-            ship_geo.TrackStation3.z,
-            ship_geo.TrackStation4.z,
-        )
-        Strawtubes.SetDeltazFrame(ship_geo.strawtubes.DeltazFrame)
-        Strawtubes.SetFrameLateralWidth(ship_geo.strawtubes.FrameLateralWidth)
-        Strawtubes.SetFrameMaterial(ship_geo.strawtubes.FrameMaterial)
-        Strawtubes.SetDeltazView(ship_geo.strawtubes.DeltazView)
-        Strawtubes.SetInnerStrawDiameter(ship_geo.strawtubes.InnerStrawDiameter)
-        Strawtubes.SetOuterStrawDiameter(ship_geo.strawtubes.OuterStrawDiameter)
-        Strawtubes.SetStrawPitch(
-            ship_geo.strawtubes.StrawPitch,
-            ship_geo.strawtubes.YLayerOffset,
-        )
-        Strawtubes.SetDeltazLayer(ship_geo.strawtubes.DeltazLayer)
-        Strawtubes.SetStrawsPerLayer(ship_geo.strawtubes.StrawsPerLayer)
-        Strawtubes.SetStereoAngle(ship_geo.strawtubes.ViewAngle)
-        Strawtubes.SetWireThickness(ship_geo.strawtubes.WireThickness)
-        Strawtubes.SetVacBox_x(ship_geo.strawtubes.VacBox_x)
-        Strawtubes.SetVacBox_y(ship_geo.strawtubes.VacBox_y)
-        Strawtubes.SetStrawLength(ship_geo.strawtubes.StrawLength)
-
-        Strawtubes.set_station_height(ship_geo.strawtubes.station_height)
-        # for the digitizing step
-        Strawtubes.SetStrawResolution(
-            ship_geo.strawtubes.v_drift,
-            ship_geo.strawtubes.sigma_spatial,
-        )
-        detectorList.append(Strawtubes)
+    configure_strawtubes(
+        os.path.join(os.environ["FAIRSHIP"], "geometry", "strawtubes_config.yaml"),
+        ship_geo,
+    )
 
     if ship_geo.EcalOption == 1:  # shashlik design TP
         ecal, EcalZSize = posEcal(ship_geo.ecal.z, ship_geo.ecal.File)